Use separate provider for isMapView state
